requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
ecdsa==0.14.1
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
install==1.3.4
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
python-dateutil==2.8.1
python-dotenv==0.17.0
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.4.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.7
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
